requirements.txt
Django==5.1
gunicorn==23.0.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1
django-cors-headers==4.3.1
whitenoise==6.6.0
djangorestframework==3.14.0
